app/requirements.txt
fastapi
uvicorn
nicegui
sqlmodel
python-dotenv
httpx
authlib
itsdangerous
fpdf2
pytest
pytest-asyncio
pytest-mock
